+++ metadata-api/src/main/java/com/hashmap/haf/metadata/config/actors/message/AbstractQueryMsg.java
@@ -1,4 +1,4 @@
-package com.hashmap.haf.metadata.config.actor.message;
+package com.hashmap.haf.metadata.config.actors.message;
 
 import com.hashmap.haf.metadata.config.model.MetadataConfigId;
 import lombok.Getter;
