+++ metadata-api/src/main/java/com/hashmap/haf/metadata/config/actors/message/metadata/DeleteMetadataConfigMsg.java
@@ -1,6 +1,6 @@
-package com.hashmap.haf.metadata.config.actor.message.metadata;
+package com.hashmap.haf.metadata.config.actors.message.metadata;
 
-import com.hashmap.haf.metadata.config.actor.message.AbstractMetadataConfigMsg;
+import com.hashmap.haf.metadata.config.actors.message.AbstractMetadataConfigMsg;
 import com.hashmap.haf.metadata.config.model.MetadataConfig;
 
 public class DeleteMetadataConfigMsg extends AbstractMetadataConfigMsg {
