+++ metadata-api/src/main/java/com/hashmap/haf/metadata/config/dao/MetadataConfigDaoImpl.java
@@ -1,6 +1,6 @@
 package com.hashmap.haf.metadata.config.dao;
 
-import com.hashmap.haf.metadata.core.common.Dao.DaoUtil;
+import com.hashmap.haf.metadata.core.common.dao.DaoUtil;
 import com.hashmap.haf.metadata.core.util.UUIDConverter;
 import com.hashmap.haf.metadata.config.entity.MetadataConfigEntity;
 import com.hashmap.haf.metadata.config.model.MetadataConfig;
